backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
strawberry-graphql[fastapi]==0.215.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
pylint==3.0.3